Convert to Bepinex and more...
Reworked Save system to use games storage and save the unique id's of the reactors and contents with charges in the save game folder.
Cached The Arial Font after first use.
Changed the Text updates for Inventory Items to only update when pda is open and the value is changed.
Removed all preprocessors as they all seem to be needing the BZ versions for both games now. (Meaning this should work on both without 2 versions.)
Some Renaming and Reorganization